Corrected renderer for cart to be smoother while moving

+++ src/main/java/com/builtbroken/icbm/content/rail/IMissileRail.java
@@ -44,5 +44,5 @@ public interface IMissileRail
      *
      * @return
      */
-    double getRailHeight();
+    float getRailHeight();
 }
